+++ src/components/StrategyEditor/tabs/StrategyTab/StrategySandboxTab.js
