+++ src/components/StrategyEditor/tabs/StrategyTab/StrategyTabWrapper.js
@@ -2,7 +2,7 @@ import React, { memo } from 'react'
 import { useSelector } from 'react-redux'
 import { getIsPaperTrading } from '../../../../redux/selectors/ui'
 import StrategyLiveTab from './StrategyLiveTab'
-import StrategySandboxTab from './StrategySandboxTabSandbox'
+import StrategySandboxTab from './StrategySandboxTab'
 
 const StrategyTabWrapper = (props) => {
   const isPaperTrading = useSelector(getIsPaperTrading)
